tests: move tests around
