+++ tests/basic.rs
@@ -1,6 +1,8 @@
 use std::str::FromStr;
 
-#[derive(enum_stringify::EnumStringify, Debug, PartialEq)]
+use enum_stringify::EnumStringify;
+
+#[derive(EnumStringify, Debug, PartialEq)]
 enum Numbers {
     One,
     Two,
@@ -42,4 +44,62 @@ fn test_from_str_trait() {
     assert_eq!(Numbers::from_str("Three").unwrap(), Numbers::Three);
 
     assert!(Numbers::from_str("Four").is_err());
+}
+
+#[derive(EnumStringify, Debug, PartialEq)]
+enum LargeEnum {
+    A1,
+    A2,
+    A3,
+    A4,
+    A5,
+    A6,
+    A7,
+    A8,
+    A9,
+    A10,
+    B1,
+    B2,
+    B3,
+    B4,
+    B5,
+    B6,
+    B7,
+    B8,
+    B9,
+    B10,
+    C1,
+    C2,
+    C3,
+    C4,
+    C5,
+    C6,
+    C7,
+    C8,
+    C9,
+    C10,
+}
+
+#[test]
+fn large_enum() {
+    assert_eq!(LargeEnum::A1.to_string(), "A1");
+    assert_eq!(LargeEnum::C10.to_string(), "C10");
+
+    assert_eq!(LargeEnum::try_from("B5").unwrap(), LargeEnum::B5);
+    assert!(LargeEnum::try_from("Z100").is_err());
+}
+
+#[derive(EnumStringify, Debug, PartialEq)]
+enum FuzzyMatch {
+    Alpha,
+    Beta,
+    Gamma,
+}
+
+#[test]
+fn wrong_similar_names() {
+    assert!(FuzzyMatch::try_from("alpha ").is_err()); // Extra space
+    assert!(FuzzyMatch::try_from("ALPHA").is_err()); // Wrong case
+    assert!(FuzzyMatch::try_from("alpHa").is_err()); // Mixed case
+    assert!(FuzzyMatch::try_from("Alphaa").is_err()); // Extra character
 }
